Simplify and reduce supported Ember configurations.
