+++ addon/templates/components/focusing-outlet.hbs
@@ -1,0 +1,1 @@
+{{#focusing-inner outletState=(-get-dynamic-var "outletState") outletName=outletName}}{{outlet outletName}}{{/focusing-inner}}
