+++ app/components/focusing-inner.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-a11y/components/focusing-inner';
