--- glimmer-version/app/components/focusing-inner.js
@@ -1,10 +1,0 @@
-import Ember from 'ember';
-
-const { Component } = Ember;
-
-export default Component.extend({
-  didUpdateAttrs() {
-    this._super(...arguments);
-    this.attrs.focus(this.get('outletState'))
-  }
-});
