--- glimmer-version/app/templates/components/focusing-outlet.hbs
@@ -1,1 +1,0 @@
-{{#focusing-inner outletState=(-get-dynamic-var "outletState") focus=(action "checkFocus")}}{{outlet outletName}}{{/focusing-inner}}
