--- glimmer-version/ember-get-owner.js
@@ -1,5 +1,0 @@
-import Ember from 'ember';
-
-const { getOwner } = Ember;
-
-export default getOwner;
