--- htmlbars-version/app/templates/components/focusing-outlet.hbs
@@ -1,3 +1,0 @@
-{{~#get-outlet-state outletName as |outletState|~}}
-  {{outlet outletName}}
-{{/get-outlet-state}}
